move String tests into separate file
